Move controls to separate component and make features configurable
